requirements.txt
# core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=0.24.0
lightfm>=1.16
pydantic>=2.0.0
faker>=8.0.0
PyYAML>=5.4.0
matplotlib>=3.4.0
seaborn>=0.11.0


# dev dependencies
pytest>=7.4.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.1
flake8>=6.1.0
pylint>=2.17.5
pre-commit>=3.3.3
